% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_attrs.R
\name{add_edge_zindex}
\alias{add_edge_zindex}
\title{Modify the edge zIndex of a sgraph object.}
\usage{
add_edge_zindex(sigma_obj, zindex)
}
\arguments{
\item{sigma_obj}{sgraph object}

\item{zindex}{Zindex value, larger is drawn above.}
}
\value{
sgraph
}
\description{
Modify the edge zIndex
}
\examples{
library(igraph)
library(sgraph)

data(lesMis)

sig <- sigma_from_igraph(igraph = lesMis) \%>\%
  add_edge_zindex(zindex = 2)
sig

}
