% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_attrs.R
\name{add_listener}
\alias{add_listener}
\title{Add a listener}
\usage{
add_listener(sigma_obj, listener)
}
\arguments{
\item{sigma_obj}{Sgraph object, created using the sigma_from_igraph function}

\item{listener}{Either "clickNode" to listen to node clicks or "hoverNode"
to listen to node hover}
}
\value{
Sgraph object with listener
}
\description{
Add a listener to report data from a 'sgraph' object in 'Shiny' back to the
R session.
}
