% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_attrs.R
\name{add_node_hidden}
\alias{add_node_hidden}
\title{Modify the node visibility of a sgraph object.}
\usage{
add_node_hidden(sigma_obj, hidden_attr)
}
\arguments{
\item{sigma_obj}{sgraph object, returned by sigma_from_igraph function}

\item{hidden_attr}{Attribute to use to set node hidden value}
}
\value{
sgraph object with modified node hidden atribute
}
\description{
Modify the node hidden attribute of an existing sgraph object.
The sgraph R package extends the sigma.js library to enable hidden nodes
that will appear upon clicking on their parent group.
}
