% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgraph_utils.R
\name{convert_to_spring_weights}
\alias{convert_to_spring_weights}
\title{Convert weights for spring layout}
\usage{
convert_to_spring_weights(df_links, selected_nodes = NULL)
}
\arguments{
\item{df_links}{Links data frame of a sgraph object}

\item{selected_nodes}{Nodes identifiers to be used for the selection.}
}
\value{
Links data frame of a sgraph object
}
\description{
Apply spring weights (revert weights).
Can add weak links to selected nodes, but best without.
}
