% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{get_color_map}
\alias{get_color_map}
\title{Build a color map}
\usage{
get_color_map(colors, palette = RColorBrewer::brewer.pal(8, "Dark2"))
}
\arguments{
\item{colors}{Groups that will be assigned to colors}

\item{palette}{Palette to use, typically a RColorBrewer palette}
}
\value{
Data frame mapping group names to colors
}
\description{
Build a color map
}
