% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{get_legend}
\alias{get_legend}
\title{Get the legend for a sgraph network}
\usage{
get_legend(colors_map, clusters)
}
\arguments{
\item{colors_map}{Color mapping to use, typically built by `get_color_map`}

\item{clusters}{Group names}
}
\value{
Ggplot object
}
\description{
Get the legend for a sgraph network
}
