% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_sgraph}
\alias{get_sgraph}
\title{Retrieve OpenDNS Security Graph data}
\usage{
get_sgraph(ip = NA, domain = NA, auth_token = NA)
}
\arguments{
\item{url}{URL of the location for the tld name authority}
}
\description{
This function will reach out to \url{https://www.opendns.org}, and retrieve
Security Graph for a passed IP or domain name.
}
\examples{
\dontrun{
sgraph_data <- get_sgraph(domain = "www.google.com", auth_token="YOUR AUTH TOKEN HERE")
}
}

