% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgraph_fit_example.R
\name{kgraph_to_lgraph}
\alias{kgraph_to_lgraph}
\title{Kgraph fit to graph list object}
\usage{
kgraph_to_lgraph(l_fit_embeds)
}
\arguments{
\item{l_fit_embeds}{kgraph object: embeddings with cosine similarity cut-off
based on random null concept pairs and known related
concept pairs}
}
\value{
graph list object: list of nodes and links data frames
}
\description{
Example function to build a graph list object (list of nodes and links data
frames) from a kgraph object (embeddings with cosine similarity cut-off
based on random null concept pairs and known related concept pairs)
}
