% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgraph_utils.R
\name{l_graph_to_igraph}
\alias{l_graph_to_igraph}
\title{Graph list object to igraph object}
\usage{
l_graph_to_igraph(l_graph)
}
\arguments{
\item{l_graph}{graph list object: list of nodes and links data frames}
}
\value{
igraph object
}
\description{
Build an igraph object from a graph list object (list of nodes and links
data frames)
}
