% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_attrs.R
\name{multiline_labels}
\alias{multiline_labels}
\title{Format multiline labels}
\usage{
multiline_labels(
  df_nodes,
  display_val_str = "\\nP-value: ",
  replace_codes = TRUE,
  label_str = "Label: ",
  group_str = "Group: "
)
}
\arguments{
\item{df_nodes}{Nodes data frame of a sgraph object}

\item{display_val_str}{String that will be prepended to each edge weight}

\item{replace_codes}{Should the label of the node replace the id ?}

\item{label_str}{String that will pe prepended to the node label}

\item{group_str}{String that will pe prepended to the node group}
}
\value{
sgraph object with modified node hidden atribute
}
\description{
The sgraph R package extends the sigma.js library to enable multiline labels.
The Javascript functions will start new lines on line breaks ('\\n') and
this function enables to easily format the details of nodes by showing the
value of the relationship (edge weights) it has with other nodes.
See the Shiny examples for use cases.
}
