% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_shiny.R
\name{renderSgraph}
\alias{renderSgraph}
\title{Render a sgraph visualization in Shiny}
\usage{
renderSgraph(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that creates a sgraph visualization}

\item{env}{Defaults to parent.frame() (cf. Shiny docs)}

\item{quoted}{Defaults to FALSE (cf. Shiny docs)}
}
\value{
Htmlwidgets render object
}
\description{
Render a sgraph visualization in Shiny
}
