% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_graph.R
\name{scale_graph}
\alias{scale_graph}
\title{Scale weights}
\usage{
scale_graph(
  weights,
  exp_scale = exp(1),
  upper_bound_mult = 25,
  lower_bound_const = 5
)
}
\arguments{
\item{weights}{Either nodes or links weights vector}

\item{exp_scale}{Scale for exponential transform}

\item{upper_bound_mult}{Constant to multiply weights by after scaling.
Use to set an upper bound for weights.}

\item{lower_bound_const}{Constant to set a lower bound for weights.
All weights below will be set to lower bound.}
}
\value{
Weights vector
}
\description{
First try to linearize the weights with the best logarithmic and polynomial,
then perform exponential scaling and set upper and lower bounds.
}
