% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_shiny.R
\name{sgraphOutput}
\alias{sgraphOutput}
\title{Create a UI element for a sgraph visualization in Shiny}
\usage{
sgraphOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{ID of the UI element}

\item{width}{Width of the UI element}

\item{height}{Height of the UI element}
}
\value{
Htmlwidgets output object
}
\description{
Create a UI element for a sgraph visualization in Shiny
}
