% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgraph_utils.R
\name{scale_graph}
\alias{scale_graph}
\title{Scale weights}
\usage{
scale_graph(
  weights,
  upper_bound_mult = if (length(weights) > 1000) 2 else 4,
  lower_bound_const = if (length(weights) > 1000) 2 else 3
)
}
\arguments{
\item{weights}{Either nodes or links weights vector}

\item{upper_bound_mult}{Constant to multiply weights by after scaling.
Use to set an upper bound for weights.}

\item{lower_bound_const}{Constant to set a lower bound for weights.
All weights below will be set to lower bound.}
}
\value{
Weights vector
}
\description{
Use non-centered scaling, multiplied by a constant, and set a lower bound
constant.
}
